Move configuration of each route into itself.
