+++ src/routes/health_check.rs
@@ -1,5 +1,9 @@
-use actix_web::HttpResponse;
+use actix_web::{web, HttpResponse};
 
-pub async fn health_check() -> HttpResponse {
+pub fn configure(cfg: &mut web::ServiceConfig) {
+    cfg.route("/health_check", web::get().to(health_check));
+}
+
+async fn health_check() -> HttpResponse {
     HttpResponse::Ok().finish()
 }
