+++ src/routes/mod.rs
@@ -1,5 +1,6 @@
-mod api;
-mod health_check;
+pub mod graphql;
+pub mod health_check;
+pub mod sw_graphql;
 
-pub use api::*;
-pub use health_check::*;
+//pub use graphql::*;
+//pub use health_check::*;
